# swagger.yaml
swagger: '2.0'
info:
  description: Arthur Clive Web Api
  version: 1.0.0
  title: Arthur Clive Swagger
  contact:
    email: raguvarthan.n@turingminds.com
host: arhturclive.swagger.io
basePath: /v1
...
schemes:
  - http
paths:
...
  '/api/user/userinfo/{username}':
    post:
      tags:
        - User
      operationId: postUserinfo
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: path
          name: username
          description: UserName of user whoes address need to be added
          required: true
          type: string
        - in: body
          name: data
          description: Object that need to be added to the DB
          required: true
          schema:
            $ref: '#/definitions/UserInfoList'
      responses:
        '200':
          description: Inserted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
    get:
      tags:
        - User
      operationId: getDefaultAddress
      produces:
        - application/xml
        - application/json
      parameters:
        - name: username
          in: path
          description: UserName of uswer whos default address details needs to be taken.
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ResponceData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
...
  '/api/user/cart/{username}':
    post:
      tags:
        - Cart
      operationId: insertProductToCart
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: path
          name: username
          description: Username of the user
          required: true
          type: string
        - in: body
          name: cartProduct
          description: Object that need to be added to the DB
          required: true
          schema:
            $ref: '#/definitions/CartList'
      responses:
        '200':
          description: Inserted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
    get:
      tags:
        - Cart
      summary: Get all the products in cart
      description: ''
      operationId: getProductsInCart
      produces:
        - application/xml
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/CartList'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
  '/api/user/wishlist/{username}':
    post:
      tags:
        - WishList
      operationId: insertProductToWishList
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: path
          name: username
          description: Username of the user
          required: true
          type: string
        - in: body
          name: wishlistProduct
          description: Object that need to be added to the DB
          required: true
          schema:
            $ref: '#/definitions/WishlistList'
      responses:
        '200':
          description: Inserted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
    get:
      tags:
        - WishList
      operationId: getProductsInWishlist
      produces:
        - application/xml
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/WishlistList'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
...
definitions:
  ResponseData:
    type: object
    properties:
      Code:
        type: number
      Message:
        type: string
      Data:
        type: object
      Content:
        type: object
...
  CartDetails:
    type: object
    properties:
      Id:
        type: number
      UserName:
        type: string
      ProductSKU:
        type: string
      MinioObject_Url:
        type: string
      ProductFor:
        type: string
      ProductType:
        type: string
      ProductDesign:
        type: string
      ProductBrand:
        type: string
      ProductPrice:
        type: number
      ProductDiscount:
        type: number
      ProductDiscountPrice:
        type: number
      ProductQuantity:
        type: number
      ProdyctSize:
        type: string
      ProductColour:
        type: string
      ProductDescription:
        type: string
  WishListDetails:
    type: object
    properties:
      Id:
        type: number
      UserName:
        type: string
      ProductSKU:
        type: string
      MinioObject_Url:
        type: string
      ProductFor:
        type: string
      ProductType:
        type: string
      ProductDesign:
        type: string
      ProductBrand:
        type: string
      ProductPrice:
        type: number
      ProductDiscount:
        type: number
      ProductDiscountPrice:
        type: number
      ProdyctSize:
        type: string
      ProductColour:
        type: string
      ProductDescription:
        type: string
  UserInfoList:
    type: object
    properties:
      ListOfAddress:
        type: array
        items:
          $ref: '#/definitions/Address'
  Address:
    type: object
    properties:
      UserName:
        type: string
      Name:
        type: string
      PhoneNumber:
        type: number
      AddressLines:
        type: string
      PostOffice:
        type: string
      City:
        type: string
      State:
        type: string
      PinCode:
        type: string
      Landmark:
        type: string
      BillingAddress:
        type: boolean
      ShippingAddress:
        type: boolean
      Default:
        type: boolean
  CartList:
    type: object
    properties:
      ListOfProducts:
        type: array
        items:
          $ref: '#/definitions/CartDetails'
  WishlistList:
    type: object
    properties:
      ListOfProducts:
        type: array
        items:
          $ref: '#/definitions/WishListDetails'
...
  ResponceData:
    type: object
    properties:
      Code:
        type: string
      Message:
        type: string
      Data:
        type: object
      Content:
        type: object
...